+++ spec/theme/column_spec.rb
@@ -2,6 +2,9 @@ require 'spec_helper'
 
 describe 'Theme', type: :request do
   describe 'Column' do
+    before(:all) do
+      @npw = NobelPrizeWinner.select('nobel_prize_winners.*, nobel_prizes.category, nobel_prizes.year').joins(:nobel_prizes)
+    end
     it 'should accept :header_attrs as an option' do
       visit '/themes/column_header_attrs'
       page.should have_xpath('//th[4][@style="background-color: red;"]')
@@ -9,8 +12,14 @@ describe 'Theme', type: :request do
 
     it 'should accept :cell_attrs as an option' do
       visit '/themes/column_cell_attrs'
-      NobelPrizeWinner.count.times do |i|
+      @npw.count.times do |i|
         page.should have_xpath("//tr[#{i+1}]/td[4][@style=\"background-color: red;\"]")
+      end
+    end
+    it 'should allow :cell_attrs to take a Proc' do
+      visit '/themes/column_cell_attrs_with_proc'
+      @npw.each_with_index do |n, i|
+        page.should have_xpath("//tr[#{i + 1}]/td[5][@style=\"color: #{n.year > 1950 ? 'green' : 'red'};\"]")
       end
     end