+++ changelog.markdown
@@ -1,3 +1,5 @@
+## 0.4.0
+
 ## 0.3.0
  * tables no longer take the :html option, all non-recognized options are passed through as HTML attributes
  * The `options` attribute on a variety of objects has been removed. It has been replaced with specialized functions such as `label_options` and `collection` with all unrecognized options placed into the `attrs` attribute.